# Conflicts:
#	source/InjectHooksMain.cpp
#	source/game_sa/SpecialFX.cpp
#	source/game_sa/VehicleRecording.cpp
#	source/game_sa/World.cpp
#	source/toolsmenu/CDebugMenu.cpp

+++ source/game_sa/OnscreenTimerEntry.cpp
@@ -19,7 +19,7 @@ void COnscreenTimerEntry::Process() {
         return;
 
     auto& timerPtr  = *CTheScripts::GetPointerToScriptVariable(m_nVarId);
-    auto  deltaTime = CTimer::GetTimeStepInMS();
+    auto deltaTime = (uint32)CTimer::GetTimeStepInMS();
 
     switch (m_nTimerDirection) {
     case eTimerDirection::INCREASE:
